refactor(trusty-common): split embedder/mod.rs into submodules under 500-SLOC cap (#607)
Splits 753-SLOC crates/trusty-common/src/embedder/mod.rs into three
focused siblings while keeping mod.rs as the thin re-export + test facade:

- types.rs         (~195 SLOC): constants, CudaOptions, ExecutionProvider,
                                Embedder trait, embed_one, resolve_* helpers,
                                is_zero_vector guard
- fast_embedder.rs (~240 SLOC): FastEmbedder struct, init_options (EP
                                selection), with_cache_size, Embedder impl
- mock.rs          (~35 SLOC):  MockEmbedder test double
- mod.rs           (~342 SLOC): candle/rss sub-mods, pub use re-exports,
                                all tests (test file classification)

Zero API surface change: all public symbols remain reachable via the
same `trusty_common::embedder::*` paths.
