+++ .line-cap-allowlist.tsv
@@ -15,7 +15,6 @@ crates/trusty-analyze/src/lang/adapters/ruby.rs	545
 crates/trusty-analyze/src/lang/adapters/typescript.rs	560
 crates/trusty-analyze/src/main.rs	567
 crates/trusty-analyze/src/mcp/mod.rs	892
-crates/trusty-common/src/embedder/mod.rs	753
 crates/trusty-common/src/memory_core/dream.rs	1199
 crates/trusty-common/src/memory_core/retrieval/mod.rs	830
 crates/trusty-common/src/memory_core/semantic_consolidation.rs	615
